refactor(ivy): remove dependency on global state from isCreationMode (#31959)
PR Close #31959
