+++ packages/core/src/render3/pure_function.ts
@@ -7,8 +7,8 @@
  */
 
 import {bindingUpdated, bindingUpdated2, bindingUpdated3, bindingUpdated4, getBinding, updateBinding} from './bindings';
-import {getBindingRoot, getLView, isCreationMode} from './state';
-
+import {getBindingRoot, getLView} from './state';
+import {isCreationMode} from './util/view_utils';
 
 
 /**
@@ -44,7 +44,7 @@ export function ɵɵpureFunction0<T>(slotOffset: number, pureFn: () => T, thisAr
   // TODO(kara): use bindingRoot instead of bindingStartIndex when implementing host bindings
   const bindingIndex = getBindingRoot() + slotOffset;
   const lView = getLView();
-  return isCreationMode() ?
+  return isCreationMode(lView) ?
       updateBinding(lView, bindingIndex, thisArg ? pureFn.call(thisArg) : pureFn()) :
       getBinding(lView, bindingIndex);
 }
